cleanup and initial setup for RecipeReview
