+++ src/sass/base/_variables.scss
@@ -6,6 +6,7 @@ $primary: #2ecc71;
 $primary-dark: #218c74;
 $secondary: #34495e;
 $info: #3498db;
+$spacer: 1rem;
 
 .text-primarydark {
     color: $primary-dark;
