+++ src/sass/components/_RecipeForm.scss
@@ -1,15 +1,3 @@
-@include media-breakpoint-up(xs) {}
-
-@include media-breakpoint-up(sm) {}
-
-@include media-breakpoint-up(md) {}
-
-@include media-breakpoint-up(lg) {}
-
-@include media-breakpoint-up(xl) {}
-
-$spacer: 1rem;
-
 // Example usage:
 @include media-breakpoint-up(xs) {
   .formLayout {
