+++ src/sass/components/_RecipeReview.scss
@@ -1,0 +1,10 @@
+.reviewPaper {
+  margin: ($spacer * 0.75) 0;
+}
+
+.empty {
+  height: 70px;
+  text-align: center;
+  vertical-align: middle;
+  line-height: 35px;
+}
