+++ src/sass/main.scss
@@ -1,1 +1,1 @@
-@import "../../node_modules/bootstrap/scss/bootstrap.scss", "./base/variables", "./base/base", "layout/home", "layout/http_404.scss", "./components/profile", "./components/Feed", "./components/RecipeCard", "./components/RecipeForm"
+@import "../../node_modules/bootstrap/scss/bootstrap.scss", "./base/variables", "./base/base", "layout/home", "layout/http_404.scss", "./components/profile", "./components/Feed", "./components/RecipeCard", "./components/RecipeForm", "./components/RecipeReview"
